repositories/flask_api/requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
PyJWT==2.9.0
Faker==30.3.0
cryptography==43.0.3